# docs/apispec.json
{
  "openapi": "3.0.3",
  "info": {
    "title": "Contact Api",
    "version": "1.0",
    "description": "API SPEC for Contact Management",
    "contact": {
      "name": "Tubagus Dylan Rachmat",
      "email": "work.tubagusdylan@gmail.com"
    },
    "license": {
      "name": "MAVEN"
    }
  },
  "servers": [
    {
      "url": "http://localhost:8080",
      "description": "Server contact management API",
      "variables": {
        "environtment": {
          "default": "dev",
          "description": "Development Server"
        }
      }
    }
  ],
  "paths": {
...
    "/api/contacts/{contactId}/addresses": {
      "post": {
        "tags": ["Address"],
        "summary": "Create new address",
        "description": "Create new address for contact",
        "parameters": [
          {
            "$ref": "#/components/parameters/HeaderToken"
          },
          {
            "$ref": "#/components/parameters/ContactId"
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/BodyAddress"
              },
              "examples": {
                "test": {
                  "value": {
                    "street": "jalan",
                    "city": "kota",
                    "province": "provinsi",
                    "country": "negara",
                    "postalCode": "10101"
                  }
                }
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Success",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Response"
                },
                "examples": {
                  "success": {
                    "value": {
                      "code": 200,
                      "message": "OK",
                      "data": {
                        "id": "address-alskdjalksjd",
                        "street": "jalan",
                        "city": "kota",
                        "province": "provinsi",
                        "country": "negara",
                        "postalCode": "10101"
                      }
                    }
                  }
                }
              }
            }
          },
          "400": {
            "description": "Validation error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Response"
                },
                "examples": {
                  "failed": {
                    "description": "Input Wrong",
                    "value": {
                      "code": 400,
                      "message": "Input wrong, ..."
                    }
                  }
                }
              }
            }
          }
        }
      },
      "get": {
        "tags": ["Address"],
        "summary": "Get list address",
        "description": "Get list address for contact",
        "parameters": [
          {
            "$ref": "#/components/parameters/HeaderToken"
          }
        ],
        "responses": {
          "200": {
            "description": "Success",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Response"
                },
                "examples": {
                  "success": {
                    "value": {
                      "code": 200,
                      "message": "OK",
                      "data": [
                        {
                          "id": "address-alskdjalksjd",
                          "street": "jalan",
                          "city": "kota",
                          "province": "provinsi",
                          "country": "negara",
                          "postalCode": "10101"
                        }
                      ]
                    }
                  }
                }
              }
            }
          },
          "404": {
            "description": "Not found error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Response"
                },
                "examples": {
                  "failed": {
                    "value": {
                      "code": 404,
                      "message": "Contact is not found"
                    }
                  }
                }
              }
            }
          }
        }
      }
    },
    "/api/contacts/{contactId}/addresses/{addressId}": {
      "get": {
        "tags": ["Address"],
        "summary": "Get detail address",
        "description": "Get detail address for contact",
        "parameters": [
          {
            "$ref": "#/components/parameters/HeaderToken"
          },
          {
            "$ref": "#/components/parameters/ContactId"
          },
          {
            "$ref": "#/components/parameters/AddressId"
          }
        ],
        "responses": {
          "200": {
            "description": "Success",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Response"
                },
                "examples": {
                  "success": {
                    "value": {
                      "code": 200,
                      "message": "OK",
                      "data": {
                        "id": "address-alskdjalksjd",
                        "street": "jalan",
                        "city": "kota",
                        "province": "provinsi",
                        "country": "negara",
                        "postalCode": "10101"
                      }
                    }
                  }
                }
              }
            }
          },
          "404": {
            "description": "Not found error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Response"
                },
                "examples": {
                  "failed": {
                    "description": "Not found",
                    "value": {
                      "code": 404,
                      "message": "Address not found"
                    }
                  }
                }
              }
            }
          }
        }
      },
      "put": {
        "tags": ["Address"],
        "summary": "Update address",
        "description": "Update address for contact",
        "parameters": [
          {
            "$ref": "#/components/parameters/HeaderToken"
          },
          {
            "$ref": "#/components/parameters/ContactId"
          },
          {
            "$ref": "#/components/parameters/AddressId"
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/BodyAddress"
              },
              "examples": {
                "test": {
                  "value": {
                    "street": "jalan",
                    "city": "kota",
                    "province": "provinsi",
                    "country": "negara",
                    "postalCode": "10101"
                  }
                }
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Success",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Response"
                },
                "examples": {
                  "success": {
                    "value": {
                      "code": 200,
                      "message": "OK",
                      "data": {
                        "id": "address-alskdjalksjd",
                        "street": "jalan",
                        "city": "kota",
                        "province": "provinsi",
                        "country": "negara",
                        "postalCode": "10101"
                      }
                    }
                  }
                }
              }
            }
          },
          "404": {
            "description": "Not found error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Response"
                },
                "examples": {
                  "failed": {
                    "description": "Not found",
                    "value": {
                      "code": 404,
                      "message": "Address not found"
                    }
                  }
                }
              }
            }
          }
        }
      },
      "delete": {
        "tags": ["Address"],
        "summary": "Delete address",
        "description": "Delete address for contact",
        "parameters": [
          {
            "$ref": "#/components/parameters/HeaderToken"
          },
          {
            "$ref": "#/components/parameters/ContactId"
          },
          {
            "$ref": "#/components/parameters/AddressId"
          }
        ],
        "responses": {
          "200": {
            "description": "Success",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Response"
                },
                "examples": {
                  "success": {
                    "value": {
                      "code": 200,
                      "message": "OK"
                    }
                  }
                }
              }
            }
          },
          "404": {
            "description": "Not found",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Response"
                },
                "examples": {
                  "success": {
                    "value": {
                      "code": 404,
                      "message": "Address not found"
                    }
                  }
                }
              }
            }
          }
        }
      }
    }
  },
  "components": {
    "parameters": {
      "HeaderToken": {
        "name": "X-API-TOKEN",
        "description": "Token for authorization",
        "required": true,
        "in": "header",
        "schema": {
          "type": "string"
        },
        "examples": {
          "token": {
            "description": "Sample token",
            "value": "asldaoichnalksjfoiqwoiehqwelkj"
          }
        }
      },
...
      "ContactId": {
        "name": "contactId",
        "description": "Contact id for get and update data",
        "required": true,
        "in": "path",
        "schema": {
          "type": "string"
        },
        "examples": {
          "user": {
            "description": "Sample contact id",
            "value": "contact-asldkjasld"
          }
        }
      },
      "AddressId": {
        "name": "addressId",
        "description": "Address id for get and update data",
        "required": true,
        "in": "path",
        "schema": {
          "type": "string"
        },
        "examples": {
          "user": {
            "description": "Sample address id",
            "value": "address-asldkjasld"
          }
        }
      }
    },
    "schemas": {
...
      "BodyAddress": {
        "type": "object",
        "properties": {
          "street": {
            "type": "string"
          },
          "city": {
            "type": "string"
          },
          "province": {
            "type": "string"
          },
          "country": {
            "type": "string"
          },
          "postalCode": {
            "type": "string"
          }
        }
      },
      "Response": {
        "type": "object",
        "properties": {
          "code": {
            "type": "number"
          },
          "message": {
            "type": "string"
          },
          "data": {
            "type": "object"
          }
        }
      },
...
    }
  }
}